+++ user_model.go
@@ -12,6 +12,12 @@ type User struct {
 	updatedAt   string
 }
 
+type UserInfo struct {
+	Id       string `json:"userID"`
+	Name     string `json:"name"`
+	SumPower string `json:"sumPower"`
+}
+
 func insertUser(user User) error {
 	const createSql = `INSERT INTO users (name, digest_token) VALUES ($1, $2);`
 	if _, err := db.Exec(createSql, user.name, user.digestToken); err != nil {
@@ -59,4 +65,29 @@ func updateUser(user User) error {
 		return fmt.Errorf("db.Exec failed: %w", err)
 	}
 	return nil
+}
+
+func selectUserRanking() ([]UserInfo, error) {
+	var users []UserInfo
+	const selectSql = `
+SELECT U.id, U.name, SUM(UOC.level * C.power) AS sumPower
+FROM user_ownership_characters AS UOC
+INNER JOIN users AS U ON UOC.user_id = U.id
+INNER JOIN characters AS C ON UOC.character_id = C.id
+GROUP BY U.id
+ORDER BY sumPower DESC
+LIMIT 3
+`
+	rows, err := db.Query(selectSql)
+	if err != nil {
+		return nil, fmt.Errorf("db.Query faild: %w", err)
+	}
+	for rows.Next() {
+		var user UserInfo
+		if err := rows.Scan(&user.Id, &user.Name, &user.SumPower); err != nil {
+			return nil, fmt.Errorf("rows.Scan faild: %w", err)
+		}
+		users = append(users, user)
+	}
+	return users, nil
 }
